requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-swagger==1.0.16
aiosignal==1.3.1
async-lru==1.0.3
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.4
dill==0.3.6
distro==1.8.0
frozenlist==1.3.3
genie==22.11
genie.libs.clean==22.11
genie.libs.conf==22.11
genie.libs.filetransferutils==22.11
genie.libs.health==22.11
genie.libs.ops==22.11
genie.libs.parser==22.11
genie.libs.sdk==22.11
gitdb==4.0.10
GitPython==3.1.18
grpcio==1.51.1
idna==3.4
Jinja2==3.1.2
jsonpickle==3.0.0
junit-xml==1.9
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.3
ncclient==0.6.13
netaddr==0.8.0
packaging==22.0
paramiko==2.12.0
pathspec==0.10.3
prettytable==3.5.0
protobuf==3.20.3
psutil==5.9.4
pyats==22.11
pyats.aereport==22.11
pyats.aetest==22.11
pyats.async==22.11
pyats.connections==22.11
pyats.datastructures==22.11
pyats.easypy==22.11
pyats.kleenex==22.11
pyats.log==22.11
pyats.reporter==22.11
pyats.results==22.11
pyats.tcl==22.11
pyats.topology==22.11
pyats.utils==22.11
pycparser==2.21
pyftpdlib==1.5.7
PyNaCl==1.5.0
python-dotenv==0.21.0
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
tftpy==0.8.0
tqdm==4.64.1
unicon==22.11
unicon.plugins==22.11
urllib3==1.26.13
wcwidth==0.2.5
xmltodict==0.13.0
yamllint==1.28.0
yang.connector==22.11
yarl==1.8.2
